refactor and fix mobile view for app bar
