+++ package.json
@@ -15,6 +15,7 @@
     "@emotion/react": "^11.13.0",
     "@emotion/styled": "^11.13.0",
     "@mui/icons-material": "^5.16.4",
+    "@mui/lab": "^5.0.0-alpha.173",
     "@mui/material": "^5.16.4",
     "@tanstack/react-form": "^0.26.4",
     "@tanstack/react-query": "^5.51.11",
